requirements.txt
alembic==1.13.1
asyncpg==0.29.0
psycopg-binary==3.1.10
SQLAlchemy==2.0.29
starlette==0.37.2
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.2
pydantic==2.7.1
pydantic-settings==2.2.1
uvicorn==0.29.0
psycopg==3.1.18
python-jose==3.3.0
aiobotocore==2.13.0
botocore==1.34.106
bcrypt==4.0.1
cryptography==44.0.0